try to use more sram than rom for descriptors/const
