+++ .github/workflows/build.yml
@@ -77,7 +77,7 @@ jobs:
   # Build ESP
   # ---------------------------------------
   esp:
-    if: false
+    #if: false
     needs: set-matrix
     uses: ./.github/workflows/build_util.yml
     with:
