--- ports/ch32v20x/board_flash.c
@@ -1,77 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2018 Ha Thach for Adafruit Industries
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-#include "board_api.h"
-
-#ifndef BUILD_NO_TINYUSB
-#include "tusb.h"
-#endif
-
-// symbol from linker
-extern uint32_t __flash_size[];
-
-//--------------------------------------------------------------------+
-//
-//--------------------------------------------------------------------+
-void board_flash_init(void) {
-
-}
-
-uint32_t board_flash_size(void) {
-  uint32_t flash_size = (uint32_t) __flash_size;
-  TU_LOG1("board_flash_size: %lu\r\n", flash_size);
-  return flash_size;
-}
-
-void board_flash_read(uint32_t addr, void* buffer, uint32_t len) {
-  (void) addr;
-  (void) buffer;
-  (void) len;
-}
-
-void board_flash_flush(void) {
-}
-
-void board_flash_write(uint32_t addr, void const* data, uint32_t len) {
-  (void) addr;
-  (void) data;
-  (void) len;
-}
-
-void board_flash_erase_app(void) {
-  // TODO implement later
-}
-
-bool board_flash_protect_bootloader(bool protect) {
-  // TODO implement later
-  (void) protect;
-  return false;
-}
-
-#ifndef BUILD_NO_TINYUSB
-void board_self_update(const uint8_t * bootloader_bin, uint32_t bootloader_len) {
-  (void) bootloader_bin;
-  (void) bootloader_len;
-}
-#endif
