+++ ports/ch32v20x/boards.c
@@ -29,6 +29,10 @@
 #include "tusb.h"
 #endif
 
+// symbol from linker
+extern uint32_t __flash_size[];
+extern uint32_t __flash_boot_size[];
+
 //--------------------------------------------------------------------+
 // MACRO TYPEDEF CONSTANT ENUM DECLARATION
 //--------------------------------------------------------------------+
@@ -103,12 +107,12 @@ void board_dfu_init(void) {
 }
 
 void board_reset(void) {
-  // NVIC_SystemReset();
+   NVIC_SystemReset();
 }
 
 void board_dfu_complete(void) {
   // Mostly reset
-  // NVIC_SystemReset();
+ NVIC_SystemReset();
 }
 
 bool board_app_valid(void) {
@@ -167,6 +171,53 @@ int board_uart_write(void const* buf, int len) {
   return 0;
 #endif
 }
+
+//--------------------------------------------------------------------+
+// Flash
+//--------------------------------------------------------------------+
+void board_flash_init(void) {
+
+}
+
+uint32_t board_flash_size(void) {
+  return (uint32_t) __flash_size;
+}
+
+void board_flash_read(uint32_t addr, void* buffer, uint32_t len) {
+  (void) addr;
+  (void) buffer;
+  (void) len;
+}
+
+void board_flash_flush(void) {
+}
+
+void board_flash_write(uint32_t addr, void const* data, uint32_t len) {
+  (void) addr;
+  (void) data;
+  (void) len;
+}
+
+void board_flash_erase_app(void) {
+  // TODO implement later
+}
+
+bool board_flash_protect_bootloader(bool protect) {
+  // TODO implement later
+  (void) protect;
+  return false;
+}
+
+#ifdef TINYUF2_SELF_UPDATE
+void board_self_update(const uint8_t * bootloader_bin, uint32_t bootloader_len) {
+  (void) bootloader_bin;
+  (void) bootloader_len;
+}
+#endif
+
+//--------------------------------------------------------------------+
+// USB Interrupt Handler
+//--------------------------------------------------------------------+
 
 #ifndef BUILD_NO_TINYUSB
 // USBFS
