+++ ports/template_port/boards.c
@@ -95,6 +95,54 @@ int board_uart_write(void const* buf, int len) {
   return 0;
 }
 
+//--------------------------------------------------------------------+
+// Flash
+//--------------------------------------------------------------------+
+void board_flash_init(void) {
+
+}
+
+uint32_t board_flash_size(void) {
+  return 0;
+}
+
+void board_flash_read(uint32_t addr, void* buffer, uint32_t len) {
+  (void) addr;
+  (void) buffer;
+  (void) len;
+}
+
+void board_flash_flush(void) {
+}
+
+void board_flash_write(uint32_t addr, void const* data, uint32_t len) {
+  (void) addr;
+  (void) data;
+  (void) len;
+}
+
+void board_flash_erase_app(void) {
+  // TODO implement later
+}
+
+bool board_flash_protect_bootloader(bool protect) {
+  // TODO implement later
+  (void) protect;
+  return false;
+}
+
+#ifdef TINYUF2_SELF_UPDATE
+void board_self_update(const uint8_t * bootloader_bin, uint32_t bootloader_len)
+{
+  (void) bootloader_bin;
+  (void) bootloader_len;
+}
+#endif
+
+//--------------------------------------------------------------------+
+// USB Interrupt Handler
+//--------------------------------------------------------------------+
+
 #ifndef BUILD_NO_TINYUSB
 // Forward USB interrupt events to TinyUSB IRQ Handler
 void OTG_FS_IRQHandler(void) {
