+++ src/usb_descriptors.c
@@ -52,7 +52,7 @@ enum {
 //--------------------------------------------------------------------+
 // Device Descriptors
 //--------------------------------------------------------------------+
-tusb_desc_device_t TINYUF2_DESC_CONST desc_device = {
+tusb_desc_device_t TINYUF2_CONST desc_device = {
     .bLength            = sizeof(tusb_desc_device_t),
     .bDescriptorType    = TUSB_DESC_DEVICE,
     .bcdUSB             = 0x0200,
@@ -104,7 +104,7 @@ uint8_t const* tud_descriptor_device_cb(void) {
   #define EPNUM_CDC_IN      0x83
 #endif
 
-uint8_t const desc_configuration[] = {
+uint8_t TINYUF2_CONST desc_configuration[] = {
     // Config number, interface count, string index, total length, attribute, power in mA
     TUD_CONFIG_DESCRIPTOR(1, ITF_NUM_TOTAL, 0, CONFIG_TOTAL_LEN, TUSB_DESC_CONFIG_ATT_REMOTE_WAKEUP, 100),
 #if CFG_TUD_CDC
